+++ Models/Owner.cs
@@ -16,7 +16,7 @@ public sealed class Owner
     public double? Rating { get; set; }
 
     public int? CompletedProjectsCount { get; set; }
-    public int? HiringRatePercent { get; set; }
+    public double? HiringRatePercent { get; set; }
 
     public string? RegisteredAt { get; set; }
     public int? OpenProjectsCount { get; set; }
